Add version to entropy
The version (default to 0) indicates how the entropy is intended to be decoded and used.

The current default version is 0. It indicates that the beacon payload is to be constructed using the sha256 of the remote entropy, remote entropy timestamp, and local entropy in that order.

This is implemented in the `entropy` workspace member of the gateway-rs repo