Add `boosted_hex_device_type_v1` for HIP-109
The first iteration of HIP-109[1] adds the ability to boost by a device
type. Pre-existing boosts will continue to boost _all_ device types.

From the way the HIP reads, at some point, a hex may be boosted by the
same provider for multiple device types. It was undecided wether that
would be implemented as multiple device_types attached to a boosted hex,
or multiple instances of a boosted hex.

[1]: https://github.com/helium/HIP/blob/main/0109-hex-boosting-by-deployment.md