add verified data transfer ingest report
The difference between valid and invalid up to this point.
- Valid :: dc has been burned for the transfer event.
- Invalid :: We don't know one of the entities involced.

The invalid report will no longer be written. It is being replaced by
Verified. All data transfer events that pass muster, we know the
entities, it's not a duplicated, etc,. are written as Verified.
Valid still means dc was burned.