update rle_event_v1 -> radio_location_correlation_v1
At this time, the most imformation we can expose about the events that
led to the radio location estimate is an id and timestamp. There may be
more information in the future, but this is enough information for
someone to give us and we can fetch the underlying record for investigation.