chain rewardable entities (#455)
* add WIP chain-rewardable-entities protos

This file represents a new data stream of entity relationships to be 
written to s3 gathered from a chain followe.

* format with clang-format

* Update rewardable entities protos

* self review

* David and Noah pair program changes

* proto 3 does not have optional values

* don't use abbreviations that could mean multiple things

re could stand for rewardable entity, or relationship depending on what
you're currently thinking about.

* keep proto version as suffixes

* report/req changes

- Keep the name of the proto written to s3 including "report"

This follows the established ingestor pattern, and makes it clearer that
what was written is what was seen in whole by the ingestor.

- Write the entire request to the bucket

The only thing the ingestor will add to the request is the time it was
seen. Otherwise, untouched.

- atomic -> update

"atomic" was trying to mean latest state, that should hopefully be clear
through comments and how the proto is used. "update" does however help
clarify that something about a hotspot changed from some source.

- Add signer and signature to update requests

These fields need to be part of the message written to the bucket so we
can go back in time and verify updates.

Added signer in the case where the identity of the notifier has to
change. Also allows us to differentiate between messages that were
correctly signed, vs messages that came from a source we trust.

* clang-format

* update => report

* Decouple updates (#462)

Co-authored-by: David Young <david@nova-labs.com>

* update => change

* block_time_seconds => ingest_time_seconds

* ingest_time_seconds => timestamp_seconds

* block_height => block

* Fix clang format issues

* Dumb formatting rules

---------

Co-authored-by: Noah Prince <chewinglass123@gmail.com>
Co-authored-by: bry <bryanzettler@gmail.com>
Co-authored-by: David Young <david@nova-labs.com>